den virkende

+++ Inc/enemies.h
@@ -46,7 +46,7 @@ void updateEnemyShoot(enemyBullet_t enemyBullet[], enemies_t[]);
 void removeEnemyShoot(enemyBullet_t enemyBullet[]);
 
 void interactionsEnemyBulletHitPlayer(enemyBullet_t enemyBullet[], spaceship_t* spaceship);
-int8_t isAllEnemyDead(enemies_t enemies[] ,spaceship_t* spaceship);
+uint16_t isAllEnemyDead(enemies_t enemies[] ,spaceship_t* spaceship);
 
 
 #endif /* ENEMIES_H_ */
